// js/stat-evaluation-player/src/generated/PositioningBallDepthEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type PositioningBallDepthEvent = { time: number, frame: number, end_time: number, end_frame: number, duration: number, player: RemoteIdTs, player_position?: [number, number, number] | null, is_team_0: boolean, behind_ball_fraction: number, level_with_ball_fraction: number, in_front_of_ball_fraction: number, };