// js/stat-evaluation-player/src/generated/PositioningBallProximityEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type PositioningBallProximityEvent = { time: number, frame: number, end_time: number, end_frame: number, duration: number, player: RemoteIdTs, player_position?: [number, number, number] | null, is_team_0: boolean, closest_to_ball_team: boolean, closest_to_ball_absolute: boolean, farthest_from_ball: boolean, };