// js/stat-evaluation-player/src/generated/PositioningFieldZoneEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type PositioningFieldZoneEvent = { time: number, frame: number, end_time: number, end_frame: number, duration: number, player: RemoteIdTs, player_position?: [number, number, number] | null, is_team_0: boolean, defensive_zone_fraction: number, neutral_zone_fraction: number, offensive_zone_fraction: number, defensive_half_fraction: number, offensive_half_fraction: number, };