// js/stat-evaluation-player/src/generated/PositioningGoalContextEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type PositioningGoalContextEvent = { time: number, frame: number, player: RemoteIdTs, player_position?: [number, number, number] | null, is_team_0: boolean, caught_ahead_of_play_on_conceded_goal: boolean, };