// js/stat-evaluation-player/src/generated/PositioningPossessionEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PositioningPossessionState } from "./PositioningPossessionState.ts";
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type PositioningPossessionEvent = { time: number, frame: number, end_time: number, end_frame: number, duration: number, player: RemoteIdTs, player_position?: [number, number, number] | null, is_team_0: boolean, possession_state: PositioningPossessionState, };