// js/stat-evaluation-player/src/generated/PositioningTeammateRoleEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PositioningTeammateRoleState } from "./PositioningTeammateRoleState.ts";
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type PositioningTeammateRoleEvent = { time: number, frame: number, end_time: number, end_frame: number, duration: number, player: RemoteIdTs, player_position?: [number, number, number] | null, is_team_0: boolean, teammate_role: PositioningTeammateRoleState, };