// js/stat-evaluation-player/src/generated/ReplayStatsPositioningSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PositioningSignalSnapshot } from "./PositioningSignalSnapshot.ts";
import type { RemoteIdTs } from "./RemoteIdTs.ts";

export type ReplayStatsPositioningSummary = { player_id: RemoteIdTs, is_team_0: boolean, distance: PositioningSignalSnapshot, };